requirements_dev.txt
-r requirements.txt
django-debug-toolbar==3.5.0
django-createsuperuserwithpassword>=2.0.0

